+++ spoolman/files/klipper/klippy/extras/spoolman/print_lifecycle.py
@@ -14,15 +14,8 @@ class PrintLifecycle:
 
     def on_print_stop(self):
         self.logs.verbose("Print job ended, clearing.")
-        self.helper.u1_tools.clear_map()
         self.helper.tracking.clear_active()
 
     def on_print_cancel(self):
         self.logs.verbose("Print job canceled, clearing.")
-        self.helper.u1_tools.clear_map()
         self.helper.tracking.clear_active()
-
-    def on_print_resume(self):
-        self.logs.verbose("Print job resumed, resuming spool tracking")
-        if self.helper.mode != "manual":
-            self.helper.u1_tools.update_map()